Add a data source for parsing questions from JSON answers dumps.
This is built up from a dump of questions that were actually uploaded to
Robotics StackExchange but should be reusable for those that weren't.

Only unanswered questions failed to migrate but it probably makes sense
to handle included answers so we can generate pages for the archived
questions in case Robotics StackExchange ever goes offline so the TODO
for handling Answers should be included.