Add nginx redirect configuration file.
This sets up an example nginx config to redirect questions which are
migrated to Robotics Stack Exchange to their new host.

The embedded config (only for ROS Answers atm) would be included in an
nginx server directive for a new static mirror of ROS Answers.

When paired with the proper server configuration it should match any
previously valid url for the original question.