Add apache httpd server image for testing config.
Our target host, OSUOSL.org will likely use httpd so that's the first
config I've tested and gotten working.
In fact I think I'll drop the nginx config because I don't think it's
going to be that useful and if we're wrong it will be easier to just
re-port from the httpd config.