Use mod_rewrite for canonical url matching.
I couldn't find a way to manage the slug -> canonical url rewriting
using only RedirectMatch without triggering a redirect loop.
There's probably a way to do it with negative lookahead but I try not to
use such tools where better tools are available and mod_rewrite with a
rewritecond is going to be easier for future me and most others to grok.